[FIX] Move utils to html_builder
